scaladoc: omit external link for type parameters
previously, the scaladoc would generate a link for type parameters in inherited definitions. this link seems useless and slightly confusing.

for example, if you have a class that extends `T => U` (Function1), the `andThen[A]` signature will link to `https://dotty.epfl.ch/api/scala/Function1.html#A-0`. i don't think this is a reference that exists on the Function1 page.

the link styling also makes the type parameter look more important than it should - it is styled like a class or trait name, rather than a mere parameter.